Refactor: Extract FancyBackground component
I've extracted the animated background with glow and grid from App.vue into a new reusable component FancyBackground.vue.

- I moved the relevant HTML structure to FancyBackground.vue.
- I moved the corresponding CSS styles to FancyBackground.vue.
- I implemented a `count` prop in FancyBackground.vue to control the number of particles, defaulting to 25.
- I integrated FancyBackground.vue into App.vue.
- I ensured `npm run non-proto-build` succeeds after changes.
